requirements.txt
streamlit==1.31.1
python-pptx==0.6.23
Pillow==10.2.0 